[fix]Controll - Service에서 DTO로 교환하기 때문에 수정

+++ football_love/src/main/java/com/deu/football_love/service/TeamServiceImpl.java
@@ -2,50 +2,96 @@ package com.deu.football_love.service;
 
 import com.deu.football_love.domain.*;
 import com.deu.football_love.domain.type.AuthorityType;
-import com.deu.football_love.dto.TeamDto;
+import com.deu.football_love.dto.*;
+import com.deu.football_love.repository.MemberRepository;
 import com.deu.football_love.repository.TeamRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.stream.Collectors;
 
-@Transactional
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class TeamServiceImpl implements TeamService{
 
     private final TeamRepository teamRepository;
+    private final MemberRepository memberRepository;
 
-    public TeamDto getTeamInfo(Team team)
+    @Transactional(readOnly = true)
+    public TeamDto getTeamInfo(String teamName)
     {
+        Team team = teamRepository.selectTeam(teamName);
+        if (team == null)
+            return null;
         return TeamDto.from(team);
     }
 
     @Override
-    public void createNewTeam(Team team) {
+    @Transactional(readOnly = false)
+    public CreateTeamResponse createNewTeam(MemberDto member, String teamName) {
+        Team team = new Team();
+        Member findMember = memberRepository.selectMember(member.getId());
+        if (findMember == null)
+            return null;
+        TeamMember teamMember = new TeamMember(team, findMember, AuthorityType.LEADER);
+        team.setName(teamName);
+        team.setCreateDate(LocalDate.now());
+        teamMember.setTeam(team);
+        teamMember.setMember(findMember);
+        team.getTeamMembers().add(teamMember);
         teamRepository.insertTeam(team);
+        teamRepository.insertNewTeamMember(teamMember);
+        return new CreateTeamResponse(teamName);
     }
 
     @Override
-    public Team findTeam(String teamName)
+    @Transactional(readOnly = true)
+    public TeamDto findTeam(String teamName)
     {
-        return teamRepository.selectTeam(teamName);
+        Team team = teamRepository.selectTeam(teamName);
+        if (team == null)
+            return null;
+        return TeamDto.from(team);
     }
 
     @Override
-    public ApplicationJoinTeam findApplication(String teamName, String memberId) {
-        return teamRepository.selectApplication(teamName, memberId);
+    @Transactional(readOnly = true)
+    public ApplicationJoinTeamDto findApplication(String teamName, String memberId) {
+        ApplicationJoinTeam application = teamRepository.selectApplication(teamName, memberId);
+        if (application == null)
+            return null;
+        return new ApplicationJoinTeamDto(application);
     }
+
     @Override
-    public void applyToTeam(ApplicationJoinTeam application) {
+    @Transactional(readOnly = false)
+    public void applyToTeam(String teamName, String memberId, String message) {
+
+        Team findTeam = teamRepository.selectTeam(teamName);
+        Member findMember = memberRepository.selectMember(memberId);
+        ApplicationJoinTeam application = new ApplicationJoinTeam(findTeam, findMember, message);
+        findTeam.getApplicationJoinTeams().add(application);
         teamRepository.insertNewApplication(application);
     }
 
     @Override
-    public void acceptApplication(ApplicationJoinTeam application, TeamMember newTeamMember) {
-        teamRepository.deleteApplication(application);
+    @Transactional(readOnly = false)
+    public AcceptApplicationResponse acceptApplication(String teamName, String memberId) {
+        ApplicationJoinTeam findApplication = teamRepository.selectApplication(teamName, memberId);
+        Team findTeam = teamRepository.selectTeam(teamName);
+        Member findMember = memberRepository.selectMember(memberId);
+        if (findApplication == null || findTeam == null || findMember == null)
+            return null;
+        teamRepository.deleteApplication(findApplication);
+        TeamMember newTeamMember = new TeamMember(findTeam, findMember, AuthorityType.MEMBER);
+        findTeam.getTeamMembers().add(newTeamMember);
         teamRepository.insertNewTeamMember(newTeamMember);
+        return new AcceptApplicationResponse(teamName, memberId);
     }
 
     /**
@@ -53,6 +99,7 @@ public class TeamServiceImpl implements TeamService{
      * 
      */
     @Override
+    @Transactional(readOnly = true)
     public AuthorityType authorityCheck(String teamName, String memberId)
     {
         List<TeamMember> teamMembers = teamRepository.selectTeamMember(teamName, memberId);
@@ -63,26 +110,45 @@ public class TeamServiceImpl implements TeamService{
     }
 
     @Override
-    public List<TeamMember> findTeamMember(String teamName, String memberId)
+    @Transactional(readOnly = true)
+    public List<TeamMemberDto> findTeamMember(String teamName, String memberId)
     {
-        return teamRepository.selectTeamMember(teamName, memberId);
+        List<TeamMember> teamMembers = teamRepository.selectTeamMember(teamName, memberId);
+        List<TeamMemberDto> collect = teamMembers.stream().map(tm -> new TeamMemberDto(tm))
+                .collect(Collectors.toList());
+        return collect;
     }
+
     @Override
+    @Transactional(readOnly = false)
     public void withdrawal(String teamName, String memberId)
     {
-        teamRepository.deleteTeamMember(teamName, memberId);
+        List<TeamMember> teamMembers = teamRepository.selectTeamMember(teamName, memberId);
+        if(teamMembers.size() != 0) {
+            teamRepository.deleteTeamMember(teamName, memberId);
+        }
     }
 
     @Override
-    public void disbandmentTeam(Team team)
+    @Transactional(readOnly = false)
+    public DisbandmentTeamResponse disbandmentTeam(String teamName)
     {
+        Team team = teamRepository.selectTeam(teamName);
+        if (team == null)
+            return null;
         teamRepository.deleteTeam(team);
+        return new DisbandmentTeamResponse(teamName);
     }
 
     @Override
-    public void updateAuthority(TeamMember member, AuthorityType authorityType)
+    @Transactional(readOnly = false)
+    public UpdateAuthorityResponse updateAuthority(String teamName, String memberId, AuthorityType authorityType)
     {
-        member.setAuthority(authorityType);
+        List<TeamMember> teamMembers = teamRepository.selectTeamMember(teamName, memberId);
+        if (teamMembers.size() == 0)
+            return null;
+        teamMembers.get(0).setAuthority(authorityType);
+        return new UpdateAuthorityResponse(teamName, memberId, authorityType);
     }
 
 }
